backend/psycho-score-backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
Pillow>=10.0.0
requests>=2.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.25.0